+++ modules/tinymce/src/plugins/accordion/main/ts/Main.ts
@@ -1,0 +1,9 @@
+import Plugin from './Plugin';
+
+Plugin();
+
+/** *****
+ * DO NOT EXPORT ANYTHING
+ *
+ * IF YOU DO ROLLUP WILL LEAVE A GLOBAL ON THE PAGE
+ *******/